requirements.txt
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
blessings==1.6
blinker==1.5
cachetools==5.3.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
command-not-found==0.3
datasets==2.8.0
dbus-python==1.2.16
decorator==5.1.1
dill==0.3.6
distro-info===0.23ubuntu1
docopt==0.6.2
entrypoints==0.4
filelock==3.9.0
flatbuffers==23.1.21
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
gpustat==0.6.0
httpie==1.0.3
huggingface-hub==0.12.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Jinja2==3.1.2
jsonschema==4.17.3
language-selector==0.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
mosec==0.6.4
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
netifaces==0.10.4
%numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py3==7.352.0
onnx==1.13.0
onnxruntime==1.14.1
onnxruntime-gpu==1.14.1
packaging==23.0
pandas==1.5.3
Pillow==9.2.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
%prometheus-client==0.16.0
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
PyGObject==3.36.0
Pympler==1.0.1
pyrsistent==0.19.3
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.17.0
sympy==1.11.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
%torch==1.13.1
tornado==6.2
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.7
ufw==0.36
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
websocket-client==1.5.1
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.2
zipp==3.12.0
#conda install pytorch==1.13.1pytorch-cuda=11.6 -c pytorch -c nvidia
#pip install prometheus-client==0.16.0
#pip install numpy==1.23.5